chore: build pow-core as a systemjs ES6 package
 - TODO: Tests are busted.
