+++ .bowerrc
@@ -1,3 +1,3 @@
 {
-  "directory": "test/vendor"
+  "directory": "vendor"
 }
