+++ .editorconfig
@@ -10,3 +10,7 @@ indent_size = 2
 
 [*.{ts,html}]
 max_line_length = 100
+
+[*.md]
+trim_trailing_whitespace = false
+insert_final_newline = false
