+++ package.json
@@ -11,8 +11,13 @@
   "license": "Apache2",
   "readmeFilename": "README.md",
   "dependencies": {
+    "underscore": "1.5.2"
+  },
+  "scripts": {
+    "test": "./node_modules/karma/bin/karma start karma.conf.js --single-run"
+  },
+  "devDependencies": {
     "grunt": "~0.4.0",
-    "underscore": "1.5.2",
     "grunt-contrib-uglify": "0.3.0",
     "grunt-typescript": "~0.7.0",
     "grunt-contrib-copy": "0.5.0",
@@ -23,12 +28,7 @@
     "karma-firefox-launcher": "0.1.3",
     "grunt-npm": "0.0.2",
     "grunt-bump": "0.0.15",
-    "grunt-conventional-changelog": "1.1.0"
-  },
-  "scripts": {
-    "test": "./node_modules/karma/bin/karma start karma.conf.js --single-run"
-  },
-  "devDependencies": {
+    "grunt-conventional-changelog": "1.1.0",
     "istanbul": "bryanforbes/istanbul#source-map-resolve",
     "karma-coverage": "gotwarlost/karma-coverage.git#source-lookup",
     "grunt-contrib-watch": "0.5.3",
@@ -38,7 +38,10 @@
     "karma-jasmine": "0.2.2",
     "karma-sauce-launcher": "0.2.14",
     "lcov-result-merger": "~1.0.2",
-    "coveralls": "~2.11.4"
+    "coveralls": "~2.11.4",
+    "grunt-systemjs-builder": "~0.2.2",
+    "dts-generator": "~1.5.0",
+    "systemjs-builder": "~0.14.7"
   },
   "contributors": []
 }
